added deeplift implementation and jupyter notebooks

+++ src/prediction/datasets/protein_sequence_dataset.py
@@ -42,4 +42,4 @@ class ProteinSequenceDataset(Dataset):
         sequence_vector = np.array([self.amino_acid_map[a] for a in sequence])
         label_vector = np.array([label])
 
-        return torch.tensor(sequence_vector, device=nn_utils.get_device()), torch.tensor(label_vector, device=nn_utils.get_device())
+        return torch.tensor(sequence_vector, device=nn_utils.get_device(), dtype=torch.float64), torch.tensor(label_vector, device=nn_utils.get_device())
